### 2019-08-23 3925663
Added the MIT License for bilateralFilter.m (http://people.csail.mit.edu/jiawen/software/LICENSE)
