### 2019-08-23 dd2ad45
Minor changes
