+++ paper/paper.bib
@@ -47,7 +47,7 @@
 
 @misc{Ramos2019,
   author = {Vítor S. Ramos},
-  title = {vitorsr/SIHR: v0.1},
+  title = {{vitorsr/SIHR: v0.1}},
   month = aug,
   year = 2019,
   doi = {10.5281/zenodo.3374929},
