### 2019-08-23 507aa4a
Struggling with pandoc
