### 2019-08-23 #1676 @kyleniemeyer
@vitorsr no worries, whedon has varying permissions—it's fine to regenerate the pdf yourself
