### 2019-08-24 #1676 @Kevin-Mattheus-Moerman
@danasolav is this something you'd be interested in reviewing? We'll get several reviewers (some of whom are experts in this field) so it is okay if this isn't exactly your expertise. You could test it for the removal of reflections from your some of your DIC data for instance.
