### 2019-08-29 854586f
Removed images. Added instructions on how to download them from archival links instead
