+++ images/README.md
@@ -1,0 +1,31 @@
+# `images`
+
+Download the zipped files from the archival links below which contain the codes for the methods of, respectively, Tan and Ikeuchi<sup>1</sup>, Shen *et al.*<sup>2</sup>, and Shen and Zheng<sup>3</sup>. Then, place the therein contained images inside this folder. See structure below.
+
+Note: the links were archived directly from the respective authors' pages (see link suffix) where their codes along with test images are freely distributed for research purposes.
+
+<sup>1</sup> http://web.archive.org/web/20190828202158/https://tanrobby.github.io/code/highlight.zip  
+<sup>2</sup> http://web.archive.org/web/20190828202109/http://ivlab.org/publications/PR2008_code.zip  
+<sup>3</sup> http://web.archive.org/web/20190828202133/http://ivlab.org/publications/AO2013_code.zip
+
+The `images` folder should have the following structure.
+
+```
+README.md
+animals.bmp
+animals_gt.bmp
+cups.bmp
+cups_gt.bmp
+fish.ppm
+fruit.bmp
+fruit_gt.bmp
+head.ppm
+lady.bmp
+masks.bmp
+masks_gt.bmp
+rabbit.bmp
+toys.ppm
+train.bmp
+watermelon.bmp
+wood.bmp
+```

[…]
