### 2019-08-29 #1676 @vitorsr
Greetings @Kevin-Mattheus-Moerman. There are a few changes to the repository that need to be made. Is it OK if I make some changes, publish a new release, and mint a new Zenodo DOI for that release while still in prereview? I'll update them accordingly. Thanks in advance.
