### 2019-08-29 #1676 @vitorsr
@kczimm @thomakra can you review this submission for JOSS?
