### 2019-09-04 #1676 @thomakra
Hi,

Apologies for the late reply.
I've been on vacation.
I should have time to have a look after Wednesday next week.
This would be my first time contributing, so I would appreciate a link to
any guidelines you have for what is expected from a review.

-Thomas-

On Wed, Sep 4, 2019, 15:19 Kevin Mattheus Moerman <notifications@github.com>
wrote:

> @wanrenjie <https://github.com/wanrenjie> @rui1996
> <https://github.com/rui1996> @ceciliavision
> <https://github.com/ceciliavision> @kczimm <https://github.com/kczimm>
> @thomakra <https://github.com/thomakra> Would you be interested in
> reviewing this submission for the Journal of Open Source Software (JOSS)?
> This package implements (MATLAB/Octave) and compares several single image
> highlight removal methods. The review process focuses
> <https://joss.readthedocs.io/en/latest/submitting.html#the-review-process>
> largely on the software <https://github.com/vitorsr/SIHR> and also a short
> paper
> <https://github.com/openjournals/joss-papers/blob/joss.01676/joss.01676/10.21105.joss.01676.pdf>.
> Let me know if you have any questions.
> This is a *prereview* issue. Once several reviewers have been found a
> review issue is opened.
>
