### 2019-09-04 #1676 @Kevin-Mattheus-Moerman
@thomakra here is information about our review process:
https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Here is an example of an ongoing review process (I'll open a similar review issue once we have sufficient reviewers), as you can see tickboxes guide reviewers through the GitHub based review process:
https://github.com/openjournals/joss-reviews/issues/1613
