### 2019-09-11 #1676 @thomakra
Here is my profile page on the university website:
https://en.uit.no/om/enhet/ansatte/person?p_document_id=138601&p_dimension_id=88136

I have a Linkedin profile, but I haven't used it actively.
My background is a M.Sc. in pattern recognition, machine learning, image
processing and remote sensing.

-Thomas-

On Wed, 11 Sep 2019 at 12:37, Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @thomakra <https://github.com/thomakra> ☝️
>
> Thanks for your reply @thomakra <https://github.com/thomakra>. Could
> please specify (or update your GitHub profile to include links to)
> personal/professional profiles and affiliations? This is so we can verify
> exactly who potential reviewers are and what their background is. Thanks.
>
