### 2019-09-16 #1676 @Kevin-Mattheus-Moerman
@wanrenjie @rui1996 @ceciliavision Would you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)? This package implements (MATLAB/Octave) and compares several single image highlight removal methods. [The review process focuses](https://joss.readthedocs.io/en/latest/submitting.html#the-review-process) largely on [the software](https://github.com/vitorsr/SIHR) and also a [short paper](https://github.com/openjournals/joss-papers/blob/joss.01676/joss.01676/10.21105.joss.01676.pdf). Let me know if you have any questions.
This is a _prereview_ issue. Once several reviewers have been found a review issue is opened.
