### 2019-10-07 #1676 @Atif-Anwer
Hi!
I would be glad to review the paper. It is quite inline with my current area of research also.
What is the deadline for submitting the review (if any)?
This would be my first time reviewing/contributing, and I have gone through the guidelines you linked above. But I'm confused as to where do reviewers submit their comments (or are they not visible to the public?)
Regards,
Atif

> @Boyiliee @Atif-Anwer @DmitryUlyanov Would you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)? This package implements (MATLAB/Octave) and compares several single image highlight removal methods. [The review process focuses](https://joss.readthedocs.io/en/latest/submitting.html#the-review-process) largely on [the software](https://github.com/vitorsr/SIHR) and also a [short paper](https://github.com/openjournals/joss-papers/blob/joss.01676/joss.01676/10.21105.joss.01676.pdf). Let me know if you have any questions.
> This is a _prereview_ issue. Once several reviewers have been found a review issue is opened.
