### 2019-10-07 #1676 @Kevin-Mattheus-Moerman
Thanks @Atif-Anwer. I aim to recruit one more reviewer and then we'll get started. In the mean time you can familiarize yourself with the [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
