### 2019-10-19 #1822 @whedon
[REVIEW]: SIHR: a MATLAB/GNU Octave toolbox for single image highlight removal
**Submitting author:** @vitorsr (<a href="http://orcid.org/0000-0002-7583-5577">Vítor Saraiva Ramos</a>)
**Repository:** <a href="https://github.com/vitorsr/SIHR" target ="_blank">https://github.com/vitorsr/SIHR</a>
**Version:** v0.1
