### 2019-10-19 #1822 @Kevin-Mattheus-Moerman
@vitorsr this is where the review takes place.

@thomakra, @Atif-Anwer, thanks for helping to review this work. The checkboxes at the top guide you through the review process. You can make comments here and/or create dedicated issues on the software's repo and link to them here.

Please let me know if you have any questions.
