### 2019-10-26 #1822 @thomakra
@Kevin-Mattheus-Moerman I've gone though the checklist. Most things seem to be in order. I've added a few issues in the repo with small suggestions for improving the documentation.
One thing I noted was that the Zenodo DOI in the paper seems to link to a removed entry and is not the same as the one linked to in the README.

I can't find any automated tests, but number of methods are small enough that I could manually test them out on the linked test images. Most seem to work fine, but I've added an issue with an Octave incompatibility, which the author is looking into.

Since this is my first review, please give me a hint if there's anything further you require from my side.
