### 2019-10-26 397534a
Round of minor changes.
(1) Updated API on Akashi2016 and Shen2013 for consistency, (2) Minor grammar improvements in the JOSS paper, (3) Fixed Zenodo DOI, (4) Fixed newline Octave compatibility, (5) Yang2010 boundary improvement, (6) Updated README.md, (7) Fixed automated testing (Octave Forge image does not have the ssim function).

This fixes #8, closes #9, and closes #10
