+++ paper/paper.bib
@@ -50,6 +50,6 @@
   title = {{vitorsr/SIHR: v0.1}},
   month = aug,
   year = 2019,
-  doi = {10.5281/zenodo.3374929},
-  url = {https://doi.org/10.5281/zenodo.3374929}
+  doi = {10.5281/zenodo.3367760},
+  url = {https://doi.org/10.5281/zenodo.3367760}
 }
