+++ Tan2005/Tan2005.m
@@ -22,7 +22,7 @@ function I_d = Tan2005(I)
 %
 %  See also SIHR, Yoon2006, Yang2010.
 
-warning(['This method is slow and noise-sensitive!', newline, ...
+warning(['This method is slow and noise-sensitive!', sprintf('\n'), ...
     '         It takes ~2 min on a small image (60+ on Octave!).'])
 
 assert(isa(I, 'float'), 'SIHR:I:notTypeSingleNorDouble', ...
