+++ utils/my_quality.m
@@ -1,3 +1,5 @@
+is_octave = exist('OCTAVE_VERSION', 'builtin'); % https://wiki.octave.org/Compatibility
+
 method = { ...
     @Tan2005, @Yoon2006, @Shen2008, @Shen2009, ...
     @Yang2010, @Shen2013, @Akashi2016};
@@ -15,7 +17,9 @@ for i = 1:length(name)
 end
 
 my_psnr = zeros(length(method), length(name));
-my_ssim = zeros(length(method), length(name));
+if ~is_octave
+    my_ssim = zeros(length(method), length(name));
+end
 
 for m = 1:length(method)
     disp(['Method ', func2str(method{m})])
@@ -23,12 +27,16 @@ for m = 1:length(method)
         disp(['  Image ', name{i}])
         I_d = feval(method{m}, image{i});
         my_psnr(m, i) = psnr(I_d, truth{i});
-        my_ssim(m, i) = ssim(I_d, truth{i});
+        if ~is_octave
+            my_ssim(m, i) = ssim(I_d, truth{i});
+        end
     end
 end
 
 disp('my_psnr =')
 disp(round(10*my_psnr)/10)
 
-disp('my_ssim =')
-disp(round(1000*my_ssim)/1000)
+if ~is_octave
+    disp('my_ssim =')
+    disp(round(1000*my_ssim)/1000)
+end
