### 2019-10-26 #1822 @vitorsr
Hi, first of all, thank you @thomakra and @Atif-Anwer for taking the time to review this submission, and @Kevin-Mattheus-Moerman for editing.

I have acted according to the feedback given by @thomakra - the [latest commit](https://github.com/vitorsr/SIHR/commit/397534a3c1c866a09dc9eba66424a4eb0aa217f4) incorporates the suggestions given and closes the corresponding issues.

Please, feel free to re-open them if any change did not apply, or create a new one for more suggestions.
