### 2019-10-29 vitorsr/SIHR#11 @Atif-Anwer
Overall comments and some QoL change recommendations
With reference to the JOSS article [review ](https://github.com/openjournals/joss-reviews/issues/1822) checklist, my overall remarks and some minor suggestions are given below:

**1.** The repository is a very welcome addition to summarize the publicly available research in the area of specular highlight mitigation, which is unfortunately quite scarce. The performance results are reproduceable in Matlab (tested in 2018b) and are thus easily verified. There are no technical issues to report, however some Quality of Life changes are recommended to increase the toolbox's ease of use for users:
- Output variable of Shen(2013) implementation should be changed to `I_d` for diffuse, keeping in line with the other scripts in the repo.
- While the performance benchmark depends on system specifications, it would be nice and informative to see a summary of the processing time of each image and method once the code has been run; similar to the PSNR and SSIM summary table generated at the end of the code run. Just by adding a `tic-toc` will be sufficient and compatible with both Matlab and Octave.
- Renaming the `my_qual.m` file to something more meaningful such as `quality_evaluation.m` or similar for a more easier and proper understanding is recommended.

**2.** The paper submitted for JOSS is well written and fulfills JOSS's requirements but the following additions will slightly improve the submission:
- A list or table of the papers referenced in the repo's README.md file should also be provided in the PDF paper for comprehensiveness and ease of (offline, PDF) reading and citation.
- Similarly, a copy of the table of the reproduced results in the paper will also make it more comprehensive and help the readers.
- in the introduction, a one liner or sample images defining what are specular highlights will also clarify the target audience of the problem tackled.

Hope you find these recommendations as helpful and beneficial to the paper.
Thanks in advance :)
