### 2019-10-29 #1822 @Atif-Anwer
@Kevin-Mattheus-Moerman Apologies for a slightly delayed review. I've reviewed the repository, the paper and gone though the checklist. Overall the paper and repository are well maintained and well written. I've added an [issue ](https://github.com/vitorsr/SIHR/issues/11#issue-513811018) in the repo with minor suggestions and QoL changes.

I successfully reproduced the claimed results in Matlab (2018b) but do not have access to or experience in Octave. However since that is already covered by some fine work by the other reviewer, it is more than satisfactory.
This is my first review for JOSS, so If the review is lacking in some places then do let me know.
Thanks and regards!
