### 2019-11-01 #1822 @Kevin-Mattheus-Moerman
@vitorsr can you comment on/deal with @Atif-Anwer 's feedback/issues?
