### 2019-11-01 #1822 @Kevin-Mattheus-Moerman
@thomakra can you review the updates implemented by @vitorsr ?
