### 2019-11-01 #1822 @vitorsr
> @vitorsr can you comment on/deal with @Atif-Anwer 's feedback/issues?

@Kevin-Mattheus-Moerman, I have noted @Atif-Anwer's concerns and am currently working on pushing changes that incorporate his feedback to the repository. Whenever I do so, I will ping you and the reviewers in this thread.

In the meantime, if @thomakra has additional comments on the updates, I'll be happy to incorporate them as well.
