### 2019-11-17 #1822 @thomakra
Sorry for not following up.
I've been moving house and I'm a little behind on a lot of things.
I'll have another look tomorrow.

-Thomas-

On Sun, Nov 17, 2019, 09:39 Kevin Mattheus Moerman <notifications@github.com>
wrote:

> @vitorsr <https://github.com/vitorsr> any updates on these developments?
>
