### 2019-11-23 6bf2a78
Closes #11. Minor changes (JOSS review); renamed variables in Shen2013 for conformity, added tic and toc for runtime metrics, also renamed my_quality.m to automated_testing.m. Moreover, I added a table with average metrics over the Shen and Zheng dataset, with references to the to-date implemented papers. Also added a brief definition of the reflection components
