+++ paper/paper.bib
@@ -1,3 +1,16 @@
+@article{Shafer1985,
+  doi = {10.1002/col.5080100409},
+  url = {https://doi.org/10.1002/col.5080100409},
+  year = {1985},
+  publisher = {Wiley},
+  volume = {10},
+  number = {4},
+  pages = {210--218},
+  author = {Steven A. Shafer},
+  title = {Using color to separate reflection components},
+  journal = {Color Research {\&} Application}
+}
+
 @incollection{Tan2014,
   doi = {10.1007/978-0-387-31439-6_538},
   url = {https://doi.org/10.1007/978-0-387-31439-6_538},
@@ -52,4 +65,108 @@
   year = 2019,
   doi = {10.5281/zenodo.3367760},
   url = {https://doi.org/10.5281/zenodo.3367760}
+}
+
+@article{Tan2005,
+  doi = {10.1109/tpami.2005.36},
+  url = {https://doi.org/10.1109/tpami.2005.36},
+  year = {2005},
+  month = feb,
+  publisher = {Institute of Electrical and Electronics Engineers ({IEEE})},
+  volume = {27},
+  number = {2},
+  pages = {178--193},
+  author = {R.T. Tan and K. Ikeuchi},
+  title = {Separating reflection components of textured surfaces using a single image},
+  journal = {{IEEE} Transactions on Pattern Analysis and Machine Intelligence}
+}
+
+@inproceedings{Yoon2006,
+  doi = {10.1109/icip.2006.312650},
+  url = {https://doi.org/10.1109/icip.2006.312650},
+  year = {2006},
+  month = oct,
+  publisher = {{IEEE}},
+  author = {Kuk-jin Yoon and Yoojin Choi and In So Kweon},
+  title = {Fast Separation of Reflection Components using a Specularity-Invariant Image Representation},
+  booktitle = {2006 International Conference on Image Processing}
+}
+
+@article{Shen2008,
+  doi = {10.1016/j.patcog.2008.01.026},
+  url = {https://doi.org/10.1016/j.patcog.2008.01.026},
+  year = {2008},
+  month = aug,
+  publisher = {Elsevier {BV}},
+  volume = {41},
+  number = {8},
+  pages = {2461--2469},
+  author = {Hui-Liang Shen and Hong-Gang Zhang and Si-Jie Shao and John H. Xin},
+  title = {Chromaticity-based separation of reflection components in a single image},
+  journal = {Pattern Recognition}
+}
+
+@article{Shen2009,
+  doi = {10.1364/ao.48.002711},
+  url = {https://doi.org/10.1364/ao.48.002711},
+  year = {2009},
+  month = may,
+  publisher = {The Optical Society},
+  volume = {48},
+  number = {14},
+  pages = {2711},
+  author = {Hui-Liang Shen and Qing-Yuan Cai},
+  title = {Simple and efficient method for specularity removal in an image},
+  journal = {Applied Optics}
+}
+
+@incollection{Yang2010,
+  doi = {10.1007/978-3-642-15561-1_7},
+  url = {https://doi.org/10.1007/978-3-642-15561-1_7},
+  year = {2010},
+  publisher = {Springer Berlin Heidelberg},
+  pages = {87--100},
+  author = {Qingxiong Yang and Shengnan Wang and Narendra Ahuja},
+  title = {Real-Time Specular Highlight Removal Using Bilateral Filtering},
+  booktitle = {Computer Vision {\textendash} {ECCV} 2010}
+}
+
+@article{Shen2013,
+  doi = {10.1364/ao.52.004483},
+  url = {https://doi.org/10.1364/ao.52.004483},
+  year = {2013},
+  month = jun,
+  publisher = {The Optical Society},
+  volume = {52},
+  number = {19},
[…]
