### 2019-11-23 #1822 @vitorsr
> @vitorsr any updates on these developments?

Yes! I have pushed changes that implemented all feedback given by @Atif-Anwer. Since it also had suggestions regarding the paper, I have re-generated a new proof.
