### 2019-11-25 #1822 @Atif-Anwer
I've gone through the updated manuscript as well as the changes committed in the repo. All the suggestions I mentioned via issue [#11](https://github.com/vitorsr/SIHR/issues/11) have been addressed to. I've also completed the review checklist above since all the issues have been verified at my end. Is there any other thing required regarding the review process from my side now? @Kevin-Mattheus-Moerman
