### 2019-11-25 #1822 @thomakra
I've had a look at the latest changes and they seem to address the provided suggestions.
I was a bit unsure about what to answer under *State of the field* as I don't think the author describes how this software compares to other commonly-used packages although the paper references relevant literature on the subject, hence I have left this unchecked.

Other than that I have no further suggestions.
