### 2019-11-25 #1822 @vitorsr
> @thomakra thanks also for your review efforts. You have a couple of boxes unchecked. Have you clarified to @vitorsr what the main issues are? Thanks.

@Kevin-Mattheus-Moerman I'd like to note that I have pushed some [changes](https://github.com/vitorsr/SIHR/commit/397534a3c1c866a09dc9eba66424a4eb0aa217f4) aimed at addressing the concerns in the issues raised by @thomakra.

> I was a bit unsure about what to answer under State of the field as I don't think the author describes how this software compares to other commonly-used packages although the paper references relevant literature on the subject, hence I have left this unchecked.

@thomakra I believe I address this in the [README.md](https://github.com/vitorsr/SIHR#raison-d%C3%AAtre) through the raison d'être (French equivalent to “statement of purpose”) subsection by linking to the source code of some of the implemented methods which were ported to MATLAB/Octave, and in the summary paper by noting that this problem has scarce software availability.

As in, there are no software packages yet on the problem except for some source code provided by some authors to reproduce their results.
