### 2019-12-16 #1822 @vitorsr
Hi @Kevin-Mattheus-Moerman, is there anything else due from my end?
