### 2019-12-26 #1822 @Kevin-Mattheus-Moerman
@vitorsr
- [ ] Your paper is about to be processed for acceptance. Can you double check the paper yourself too?
- [ ] Can you check and extend you affiliation to include the city and country?
- [ ] Can you rephrase the following `...starting from the most performing ones...`? Perhaps use something like this instead (feel free to disagree and propose an alternative form): `...starting with the most computationally efficient ones...`
- [ ] Change `As of time of writing, the following methods are available.` to `At the time of writing the methods listed in Table 1 are available.`
- [ ] The final paper will contain a link to the latest archived version. Therefore please remove this sentence: `The source code for SIHR is being archived by Zenodo since its pre-release version (Ramos, 2019).`(you can also clean up the .bib file when this reference is removed).
