### 2019-12-26 b833783
Minor changes (JOSS review)
See also openjournals/joss-reviews#1822
