+++ paper/paper.bib
@@ -58,15 +58,6 @@
   booktitle = {Lecture Notes in Computer Science}
 }
 
-@misc{Ramos2019,
-  author = {Vítor S. Ramos},
-  title = {{vitorsr/SIHR: v0.1}},
-  month = aug,
-  year = 2019,
-  doi = {10.5281/zenodo.3367760},
-  url = {https://doi.org/10.5281/zenodo.3367760}
-}
-
 @article{Tan2005,
   doi = {10.1109/tpami.2005.36},
   url = {https://doi.org/10.1109/tpami.2005.36},
