### 2019-12-26 #1822 @vitorsr
@Kevin-Mattheus-Moerman, thanks for the additions. I have incorporated all in the latest commit. Also,  I have double-checked the paper, and it is sound on my end.
