### 2020-01-11 #1822 @Kevin-Mattheus-Moerman
@vitorsr at this point can you please archive the software on Zenodo and report back with the DOI of the archived version here?
Please make sure the meta data for the archived version (i.e. the authors and title) match that of the JOSS paper.
