### 2020-01-12 #1822 @vitorsr
@Kevin-Mattheus-Moerman the version v0.1 is still accurate. The DOI of the archived version is [10.5281/zenodo.3367760](https://doi.org/10.5281/zenodo.3367760) (all versions, resolves to the latest archive.)
